// src/apis/models/FireBytes_Unified_Wave_ControlPanels_Dto_SetFloorPlanViewerConfigInput.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FireBytes_Unified_Wave_ControlPanels_Dto_SetFloorPlanViewerConfigInput = {
    isThreeD?: boolean;
    controlPanelId?: string;
    graphicPanelId?: string;
    viewerCameraConfig?: string | null;
};