// src/apis/models/FireBytes_Unified_Wave_ControlPanels_Dto_ControlPanelImageMarkerDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FireBytes_Unified_Wave_ControlPanels_Dto_ControlPanelImageMarkerDto = {
    devicePointViewId?: string;
    title?: string | null;
    deviceStatusId?: string | null;
    pointXvalue?: number;
    pointYvalue?: number;
    deviceId?: string;
    deviceType?: string | null;
    commandMode?: string | null;
    peopleAround?: Array<string> | null;
    isIndoorTrackingDevice?: boolean;
};